// packages/server-nest/src/modules/FinancialStatements/modules/BalanceSheet/BalanceSheet.types.ts
export enum BALANCE_SHEET_SCHEMA_NODE_TYPE {
  AGGREGATE = 'AGGREGATE',
  ACCOUNTS = 'ACCOUNTS',
  ACCOUNT = 'ACCOUNT',
  NET_INCOME = 'NET_INCOME',
}
// ...
export interface IBalanceSheetCommonNode {
  total: IBalanceSheetTotal;
  horizontalTotals?: IBalanceSheetTotal[];

  percentageRow?: IBalanceSheetPercentageAmount;
  percentageColumn?: IBalanceSheetPercentageAmount;

  previousPeriod?: IBalanceSheetTotal;
  previousPeriodChange?: IBalanceSheetTotal;
  previousPeriodPercentage?: IBalanceSheetPercentageAmount;

  previousYear?: IBalanceSheetTotal;
  previousYearChange?: IBalanceSheetTotal;
  previousYearPercentage?: IBalanceSheetPercentageAmount;
}

export interface IBalanceSheetAggregateNode extends IBalanceSheetCommonNode {
  id: string;
  name: string;
  nodeType: BALANCE_SHEET_SCHEMA_NODE_TYPE.AGGREGATE;
  children?: IBalanceSheetDataNode[];
}

export interface IBalanceSheetTotal {
  amount: number;
  formattedAmount: string;
  currencyCode: string;
  date?: string | Date;
}

export interface IBalanceSheetAccountsNode extends IBalanceSheetCommonNode {
  id: number | string;
  name: string;
  nodeType: BALANCE_SHEET_SCHEMA_NODE_TYPE.ACCOUNTS;
  children: IBalanceSheetAccountNode[];
}

export interface IBalanceSheetAccountNode extends IBalanceSheetCommonNode {
  id: number;
  index: number;
  name: string;
  code: string;
  parentAccountId?: number;
  nodeType: BALANCE_SHEET_SCHEMA_NODE_TYPE.ACCOUNT;
  children?: IBalanceSheetAccountNode[];
}

export interface IBalanceSheetNetIncomeNode extends IBalanceSheetCommonNode {
  id: number;
  name: string;
  nodeType: BALANCE_SHEET_SCHEMA_NODE_TYPE.NET_INCOME;
}

export type IBalanceSheetDataNode =
  | IBalanceSheetAggregateNode
  | IBalanceSheetAccountNode
  | IBalanceSheetAccountsNode
  | IBalanceSheetNetIncomeNode;

export interface IBalanceSheetPercentageAmount {
  amount: number;
  formattedAmount: string;
}

export interface IBalanceSheetSchemaAggregateNode {
  name: string;
  id: string;
  type: BALANCE_SHEET_SCHEMA_NODE_TYPE;
  children: IBalanceSheetSchemaNode[];
  alwaysShow: boolean;
}

export interface IBalanceSheetSchemaAccountNode {
  name: string;
  id: string;
  type: BALANCE_SHEET_SCHEMA_NODE_TYPE;
  accountsTypes: string[];
}

export interface IBalanceSheetSchemaNetIncomeNode {
  id: string;
  name: string;
  type: BALANCE_SHEET_SCHEMA_NODE_TYPE;
}

export type IBalanceSheetSchemaNode =
  | IBalanceSheetSchemaAccountNode
  | IBalanceSheetSchemaAggregateNode
  | IBalanceSheetSchemaNetIncomeNode;

export interface IBalanceSheetDatePeriods {
  assocAccountNodeDatePeriods(node): any;
  initDateRangeCollection(): void;
}

export interface IBalanceSheetComparsions {
  assocPreviousYearAccountNode(node);
  hasPreviousPeriod(): boolean;
  hasPreviousYear(): boolean;
  assocPreviousPeriodAccountNode(node);
}

export interface IBalanceSheetTotalPeriod extends IFinancialSheetTotalPeriod {
  percentageRow?: IBalanceSheetPercentageAmount;
  percentageColumn?: IBalanceSheetPercentageAmount;
}

export interface IFinancialSheetTotalPeriod {
  fromDate: any;
  toDate: any;
  total: any;
}